v1.2.1
fix WRITE ONCE MODE bug

+++ timeLibSub.h
@@ -2,6 +2,11 @@
 // 電池長期間駆動Arduinoロガー
 // deslemLogger (deep sleep EEPROM logger)用補助スケッチ
 // わかりやすいようにメインのスケッチから日時処理を分離
+// 2038年問題を避けるため、UNIX時間は使わない。Arduinoはlongまでしか使えない。unsigned longにしてもよいが引算がやっかい。
+// 時刻にはtmElements_t型を使っている。これはTimeLib.hで定義されている。
+// 年には1970年からの経過年数をⅠバイトで管理している。このため、1970+255=2225年まで対応
+// ただし、ところどころで西暦を2桁で扱っているので実質的に2099年までしか対応しない。
+// このプログラムがいつまで使用されるかわからないが、十分であろう。
 //////////////////////////////////////////////////////
 
 #ifndef _timeLibSub_h
@@ -9,10 +14,14 @@
 #include <TimeLib.h>            // https://github.com/PaulStoffregen/Time
 
 
+//////////////////////////////////////////////////////////////////////////////
+// 以下はヘッダからの経過時間算出用なので、余り長い間隔には対応していない。必要ならlongにする。
+//////////////////////////////////////////////////////////////////////////////
 int daysDiff(tmElements_t sTm, tmElements_t eTm);
-int hourDiff(tmElements_t sTm, tmElements_t eTm);
+long hourDiff(tmElements_t sTm, tmElements_t eTm);  // ログファイル書出し時対応のためこれだけはlong
 int minDiff(tmElements_t sTm, tmElements_t eTm);
 int secDiff(tmElements_t sTm, tmElements_t eTm);
+
 int daysInYear(int year);
 boolean timeMinuteEqual(tmElements_t tm, tmElements_t tmLast);
 int dayOfYear(int yy, byte mm, byte dd);
